split atlas-mask, brain, library
